Extract module
